Add new Climber subsystem and Climb command.
I added a climber subsystem. The motor used is undefined so far.
The port numbers for the motor, and the motor's connection to the PDP need to
be changed. In addition the stalling threshold of the motor needs to be changed after testing.

The reason the climb command has a speed parameter is that I believe the Robot
should fetch how much time is left in the match to decide how fast it should
climb the rope. So whatever calls the Climb command needs to use this.